Bulk delete api for objects
Currently only supports groups and actors.  Requires superuser access to post.
JSON body is expected to be of the form:
{
  "type": "type_name",
  "collection": ["authz_ids"]
}

Where the type_name should be either group or actor and the collection should
be an array of authz_ids.  Will return error in the event of a database error
otherwise will return ok with the response body of the form:
{
  "count": num_deleted
}

Where num_deleted is the number of authz_ids deleted.  This may not match the
length of the array in the event one of the target authz_ids was missing at
time of delete.